requirements.txt
google-generativeai
py-cord==2.4.1
python-dotenv==1.0.0
wavelink==3.1.0
